Fix: product detail 리팩토링
