+++ src/context/store.ts
@@ -11,7 +11,7 @@ interface UserState extends User {
 const useStore = create(
   devtools<UserState>(set => ({
     token: '',
-    isLogin: false,
+    isLogin: true,
     login(loginRes: User) {
       set(() => ({ ...loginRes, isLogin: true }));
     },
